refactor(test-benchmark): organize stateful benchmark folder structure (#3152)
* chore: relocate balance stateful benchmark

 - Rename test_bloatnet_balance_opcode to test_balance_query for readability
 - Move `test_balance_query` from `test_multi_opocode` to `test_account_query` folder

* refactor: rename test_create2_access to test_create_operation

* chore: organize call related stateful benchmark
 - create test_call_operation file
 - move test_call_value_existing and test_call_value_new_account
 - rename test name

* refactor: relocate erc20 elated benchmark

 - add test_erc20_operation file
 - move test_sload_erc20_generic, test_sstore_erc20_generic, test_mixed_sload_sstore to same file

* refactor: relocate sload benchmark

 - add test_sload file
 - move sload related benchmark

* refactor: relocate sstore benchmark

 - add test_sstore file
 - move test_sstore_bloated, test_sstore_dirty_transitions to new file

* chore: move test_extcodesize_bytecode_sizes to test_account_query

* chore: docstring, comment refactor

* docs: update stub-dependent test examples after benchmark reorg

* chore(tests): fix markdown rendering in bloatnet benchmark docstrings

* refactor(test-benchmark): file renaming, remove _operation suffix

---------
