+++ tests/benchmark/stateful/bloatnet/depth_benchmarks/__init__.py
@@ -1,3 +1,1 @@
-"""
-abstract: BloatNet worst-case attack benchmark for maximum SSTORE stress.
-"""
+"""BloatNet worst-case attack benchmark for maximum SSTORE stress."""
