+++ tests/benchmark/stateful/bloatnet/test_account_query.py
@@ -5,14 +5,20 @@ from typing import Any
 import pytest
 from execution_testing import (
     Account,
+    Address,
     Alloc,
     BenchmarkTestFiller,
+    Block,
+    BlockchainTestFiller,
     Bytecode,
+    Conditional,
+    Create2PreimageLayout,
     Fork,
     Hash,
     IteratingBytecode,
     JumpLoopGenerator,
     Op,
+    Storage,
     TestPhaseManager,
     Transaction,
     While,
@@ -23,7 +29,9 @@ from tests.benchmark.helper.account_creator import (
     AccountMode,
 )
 from tests.benchmark.stateful.helpers import (
+    DECREMENT_COUNTER_CONDITION,
     CacheStrategy,
+    build_benchmark_txs,
     build_cache_strategy_blocks,
 )
 
@@ -327,4 +335,286 @@ def test_account_access(
         target_opcode=target_opcode,
         skip_gas_used_validation=True,
         expected_receipt_status=1,
+    )
+
+
+@pytest.mark.stub_parametrize("factory_stub", "bloatnet_factory_")
+@pytest.mark.parametrize(
+    "second_opcode",
+    [Op.EXTCODESIZE, Op.EXTCODECOPY, Op.EXTCODEHASH, Op.STATICCALL, Op.CALL],
+)
+@pytest.mark.parametrize(
+    "balance_first",
+    [True, False],
+    ids=["balance_first", "opcode_first"],
+)
+def test_balance_query(
+    benchmark_test: BenchmarkTestFiller,
+    pre: Alloc,
+    fork: Fork,
+    gas_benchmark_value: int,
+    tx_gas_limit: int,
+    balance_first: bool,
+    second_opcode: Op,
+    factory_stub: str,
+) -> None:
+    """Benchmark BALANCE paired with a second opcode on factory contracts."""
+    factory_address = pre.deploy_contract(
+        code=Bytecode(),
+        stub=factory_stub,
+    )
+
+    # Contract Construction
+    setup = Bytecode()
+
+    setup += Conditional(
+        condition=Op.STATICCALL(
+            gas=Op.GAS,
+            address=factory_address,
+            args_offset=0,
+            args_size=0,
+            ret_offset=96,
+            ret_size=64,
+            # gas accounting
+            address_warm=False,
+            old_memory_size=0,
+            new_memory_size=160,
+        ),
+        if_false=Op.INVALID,
+    )
+
+    create2_preimage = Create2PreimageLayout(
+        factory_address=factory_address,
+        salt=Op.CALLDATALOAD(32),
+        init_code_hash=Op.MLOAD(128),
+        old_memory_size=160,
+    )
+
+    setup += create2_preimage
+    setup += Op.CALLDATALOAD(0)  # [num_contract]
+
+    # Build the second opcode's bytecode
+    balance_op = Op.POP(Op.BALANCE)
+
+    if second_opcode == Op.EXTCODESIZE:
+        other_op = Op.POP(Op.EXTCODESIZE)
+    elif second_opcode == Op.EXTCODECOPY:
+        max_contract_size = fork.max_code_size()
+        other_op = Op.POP(
+            Op.EXTCODECOPY(
+                address=Op.DUP4,
+                dest_offset=Op.ADD(Op.MLOAD(32), 96),
+                offset=max_contract_size - 1,
+                size=1,
+                data_size=1,
+            )
+        )
+    elif second_opcode == Op.EXTCODEHASH:
+        other_op = Op.POP(Op.EXTCODEHASH)
+    elif second_opcode == Op.STATICCALL:
+        # gas=1: forces account/code loading, then fails
+        other_op = (
+            Op.POP(
+                Op.STATICCALL(
+                    gas=1,
+                    address=Op.DUP5,
+                    args_offset=0,
+                    args_size=0,
+                    ret_offset=0,
+                    ret_size=0,
+                )
+            )
+            + Op.POP
+        )
+    elif second_opcode == Op.CALL:
+        # gas=1: forces account/code loading, then fails
+        other_op = (
+            Op.POP(
+                Op.CALL(
+                    gas=1,
+                    address=Op.DUP6,
+                    value=0,
+                    args_offset=0,
+                    args_size=0,
+                    ret_offset=0,
+                    ret_size=0,
+                )
+            )
+            + Op.POP
+        )
+    else:
+        raise ValueError(f"Unsupported opcode: {second_opcode}")
+
+    benchmark_ops = (
+        (balance_op + other_op) if balance_first else (other_op + balance_op)
+    )
+
+    loop = While(
+        body=(
+            create2_preimage.address_op()
+            + Op.DUP1
+            + benchmark_ops
+            + create2_preimage.increment_salt_op()
+        ),
+        condition=DECREMENT_COUNTER_CONDITION,
+    )
+
+    # Contract Deployment
+    code = setup + loop
+    attack_contract_address = pre.deploy_contract(code=code)
+
+    # Gas Accounting
+    txs, total_gas_consumed = build_benchmark_txs(
+        pre=pre,
+        fork=fork,
+        gas_benchmark_value=gas_benchmark_value,
+        tx_gas_limit=tx_gas_limit,
+        attack_contract_address=attack_contract_address,
+        setup_cost=setup.gas_cost(fork),
+        iteration_cost=loop.gas_cost(fork),
+    )
+
+    benchmark_test(
+        pre=pre,
+        blocks=[Block(txs=txs)],
+        expected_benchmark_gas_used=total_gas_consumed,
+        skip_gas_used_validation=True,
+    )
+
+
+def get_factory_stub_name(size_kb: float) -> str:
+    """Generate stub name for factory based on size."""
+    if size_kb == 0.5:
+        return "bloatnet_factory_0_5kb"
+    elif size_kb == 1.0:
+        return "bloatnet_factory_1kb"
+    elif size_kb == 2.0:
+        return "bloatnet_factory_2kb"
+    elif size_kb == 5.0:
+        return "bloatnet_factory_5kb"
+    elif size_kb == 10.0:
+        return "bloatnet_factory_10kb"
+    elif size_kb == 24.0:
+        return "bloatnet_factory_24kb"
+    else:
+        raise ValueError(f"Unsupported size: {size_kb}KB")
+
+
+def build_attack_contract(factory_address: Address) -> Bytecode:
+    """Build the EXTCODESIZE attack contract with a gas-based loop exit."""
+    gas_reserve = 50_000  # Reserve for 2x SSTORE + cleanup
+    num_deployed_offset = 96
+    init_code_hash_offset = num_deployed_offset + 32
+    return_size = 64
+    return (
+        # Call factory.getConfig() -> (num_deployed, init_code_hash)
+        Conditional(
+            condition=Op.STATICCALL(
+                gas=Op.GAS,
+                address=factory_address,
+                args_offset=0,
+                args_size=0,
+                # MEM[num_deployed_offset]=num_deployed
+                # MEM[num_deployed_offset + 32]=init_code_hash
+                ret_offset=num_deployed_offset,
+                ret_size=return_size,
+            ),
+            if_false=Op.REVERT(0, 0),
+        )
+        + (
+            create2_preimage := Create2PreimageLayout(
+                factory_address=factory_address,
+                salt=Op.SLOAD(0),
+                init_code_hash=Op.MLOAD(init_code_hash_offset),
+                old_memory_size=num_deployed_offset + return_size,
+            )
+        )
+        + Op.MSTORE(160, 0)  # Initialize last_size
+        + While(
+            body=(
+                Op.MSTORE(160, Op.EXTCODESIZE(create2_preimage.address_op()))
+                + create2_preimage.increment_salt_op()
+            ),
+            condition=(
+                Op.AND(
+                    Op.GT(Op.GAS, gas_reserve),
+                    # num_deployed > salt
+                    Op.GT(
+                        Op.MLOAD(num_deployed_offset),
+                        Op.MLOAD(create2_preimage.salt_offset),
+                    ),
+                )
+            ),
+        )
+        + Op.SSTORE(0, Op.MLOAD(32))  # Save final salt
+        + Op.SSTORE(1, Op.MLOAD(160))  # Save last result
+        + Op.STOP
+    )
+
+
+@pytest.mark.parametrize(
+    "bytecode_size_kb",
+    [0.5, 1.0, 2.0, 5.0, 10.0, 24.0],
+    ids=lambda size: f"{size}KB",
+)
+def test_extcodesize_bytecode_sizes(
+    blockchain_test: BlockchainTestFiller,
+    pre: Alloc,
+    bytecode_size_kb: float,
+    gas_benchmark_value: int,
+    tx_gas_limit: int,
+) -> None:
+    """Execute EXTCODESIZE benchmark against pre-deployed contracts."""
+    expected_size_bytes = int(bytecode_size_kb * 1024)
+
+    # Get factory stub name for this size
+    factory_stub = get_factory_stub_name(bytecode_size_kb)
+
+    # Deploy factory stub (address comes from stub file)
+    factory_address = pre.deploy_contract(
+        code=Bytecode(),  # Empty bytecode - address from stub
+        stub=factory_stub,
+    )
+
+    # Build and deploy the attack contract
+    attack_code = build_attack_contract(factory_address)
+    attack_address = pre.deploy_contract(code=attack_code)
+
+    # Calculate how many transactions we need to fill the block
+    num_attack_txs = gas_benchmark_value // tx_gas_limit
+    if num_attack_txs == 0:
+        num_attack_txs = 1
+
+    # Fund the sender
+    sender = pre.fund_eoa()
+
+    # Build transactions
+    txs = []
+
+    # Attack transactions: all identical, no calldata needed
+    for _ in range(num_attack_txs):
+        attack_tx = Transaction(
+            gas_limit=tx_gas_limit,
+            to=attack_address,
+            sender=sender,
+        )
+        txs.append(attack_tx)
+
+    # Create block with all transactions
+    block = Block(txs=txs)
+
+    # Post-state verification:
+    # Attack contract slot 1 = expected size (last EXTCODESIZE result)
+    # Slot 0 can be any value (final salt depends on gas used)
+    attack_storage = Storage({1: expected_size_bytes})  # type: ignore[dict-item]
+    attack_storage.set_expect_any(0)
+
+    post = {
+        attack_address: Account(storage=attack_storage),
+    }
+
+    blockchain_test(
+        pre=pre,
+        post=post,
+        blocks=[block],
     )
